Renamed straincollection to strain_collection
