+++ src/nplinker/strain_collection.py
